Ch 5 bug fix
I am trying to fix the bug in the resampling chapter -- "sumamriseResults" for the FIsher IFnoramtion metric.
For some reason, FI.sd == NaN for many cases where, I think, it should not be.

I am also updating the summariseResults function to plot the log(FI for better visualization) of summary results.